Algo look to works. Have to do some further test and do some optimisation.

+++ srcs/AStar/AStar.hpp
@@ -6,7 +6,7 @@
 /*   By: jdugoudr <jdugoudr@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/09 14:12:59 by jdugoudr          #+#    #+#             */
-/*   Updated: 2021/02/11 15:55:19 by jdugoudr         ###   ########.fr       */
+/*   Updated: 2021/02/12 16:15:09 by jdugoudr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,11 +36,11 @@ private:
 
 	void		isAlreadyKnown(std::list<Node*> *lst);
 	void		for_each_neighbor(Node *curr, std::list<Node*> neighbors);
-	Node							*swapMap(size_t, size_t, size_t, size_t);
+	Node							*swapMap(int, int, int, int);
 	std::list<Node *>	getNeighbor();
 	void							createNeighbor(std::list<Node *> &lst,
-																						size_t i,
-																						size_t j);
+																						int i,
+																						int j);
 
 public:
 	AStar(Node &start, Node const &goal, IHeuristic &h);
